+++ test/helpers/vfile.js
@@ -1,0 +1,5 @@
+exports.fileWithNoMessages = { source: 'retext-spell', messages: [] };
+
+exports.buildVfile = actuals => ({
+  messages: actuals.map(actual => ({ source: 'retext-spell', actual })),
+});